Move static data structs from headers to code
