+++ firmware/tinyg/settings.h
@@ -105,10 +105,10 @@
 
 // machine default profiles - choose only one:
 
-#include "settings/settings_default.h"				// Default settings for release
+//#include "settings/settings_default.h"				// Default settings for release
 //#include "settings/settings_cnc3040.h"
 //#include "settings/settings_test.h"					// Settings for testing - not for release
-//#include "settings/settings_openpnp.h"				// OpenPnP
+#include "settings/settings_openpnp.h"				// OpenPnP
 //#include "settings/settings_othermill.h"				// OMC OtherMill
 //#include "settings/settings_probotixV90.h"			// Probotix Fireball V90
 //#include "settings/settings_shapeoko2.h"				// Shapeoko2 - standard kit
