+++ firmware/tinyg/xio.c
@@ -88,6 +88,12 @@ typedef struct xioSingleton {
 } xioSingleton_t;
 xioSingleton_t xio;
 
+// Static structure allocations
+xioDev_t 		ds[XIO_DEV_COUNT];			// allocate top-level dev structs
+xioUsart_t 		us[XIO_DEV_USART_COUNT];	// USART extended IO structs
+xioSpi_t 		spi[XIO_DEV_SPI_COUNT];		// SPI extended IO structs
+xioFile_t 		fs[XIO_DEV_FILE_COUNT];		// FILE extended IO structs
+
 /********************************************************************************
  * XIO Initializations, Resets and Assertions
  */
