+++ firmware/tinyg/xio.h
@@ -163,11 +163,10 @@ typedef void (*x_flow_t)(xioDev_t *d);
 #include "xio/xio_usart.h"
 #include "xio/xio_spi.h"
 
-// Static structure allocations
-xioDev_t 		ds[XIO_DEV_COUNT];			// allocate top-level dev structs
-xioUsart_t 		us[XIO_DEV_USART_COUNT];	// USART extended IO structs
-xioSpi_t 		spi[XIO_DEV_SPI_COUNT];		// SPI extended IO structs
-xioFile_t 		fs[XIO_DEV_FILE_COUNT];		// FILE extended IO structs
+extern xioDev_t 		ds[XIO_DEV_COUNT];			// allocate top-level dev structs
+extern xioUsart_t 		us[XIO_DEV_USART_COUNT];	// USART extended IO structs
+extern xioSpi_t 		spi[XIO_DEV_SPI_COUNT];		// SPI extended IO structs
+extern xioFile_t 		fs[XIO_DEV_FILE_COUNT];		// FILE extended IO structs
 extern struct controllerSingleton tg;		// needed by init() for default source
 
 /*************************************************************************
